+++ CHANGES.md
@@ -11,6 +11,8 @@
 
 ## develop
 
+- [ADD] Encoder に再初期化なしで動的プロパティを再設定する `reconfigure` API を追加する
+  - @melpon
 
 ### misc